Simplify scripts, remove .pot files, update READMEs

+++ po4a-update-templates.sh
@@ -1,9 +1,8 @@
 #!/bin/bash
 # You need po4a > 0.54, see https://github.com/mquinson/po4a/releases
 # There is no need for system-wide installation of po4a
-# You may set following variables
+# You may set following variables:
 # SRC_DIR folder for original English .md files
-# POT_DIR folder where the .pot template files are created/updated
 # PO_DIR directory where .po files are stored
 
 
@@ -14,14 +13,9 @@
 # Folder where source English .md files are
 SRC_DIR="./wiki/en"
 
-# Folder where the .pot template files are created/updated
-if [ -z "$POT_DIR" ] ; then
-    POT_DIR="./translator-files/l10n/templates"
-fi
-
 # Directory where the po file folders are
 if [ -z "$PO_DIR" ] ; then
-	PO_DIR="./translator-files/l10n/po"
+	PO_DIR="./translator-files/po"
 fi
 
 ####################################
@@ -35,42 +29,30 @@ if ! [ -x "$(command -v po4a)" ] ; then
 fi
 
 # Check if source document folder exists in the right place
-if [ ! -d "$SRC_DIR" ] ; then
+if ! [ -d "$SRC_DIR" ] ; then
     echo Error: please run this script from the root folder. >&2
     exit 1
 fi
 
-############################################
-# CREATE/UPDATE .pot TEMPLATES and .po files
-############################################
+####################################
+# CREATE/UPDATE .po FILES
+####################################
 
-while IFS= read -r -d '' file
-do
+while IFS= read -r -d '' file ; do
     # Determine target file/folder names
     basename=$(basename -s .md "$file")
-    dirname=$(dirname "$file")
-    path="${dirname#$SRC_DIR/}"
 
-    if [ "$dirname" = "$SRC_DIR" ] ; then
-        potname="$basename.pot"
-    else
-        potname="$path/$basename.pot"
-        mkdir -p "$POT_DIR/$path"
-    fi
+    for lang in $(ls "$PO_DIR") ; do
 
-    # Use source .md files in English to update/create .pot templates and .po files
-    po4a-gettextize \
-        --format asciidoc \
-        --master "$file" \
-        --master-charset "UTF-8" \
-        --po "$POT_DIR/$potname"
-
-    for lang in $(ls "$PO_DIR" ) ; do
-
-        po_file="$PO_DIR/$lang/${potname%.pot}.po"
+        po_file="$PO_DIR/$lang/${basename}.po"
 
         # po4a-updatepo will complain if the following is not met
         sed -i 's/Content-Type: text\/plain; charset=CHARSET/Content-Type: text\/plain; charset=UTF-8/g' "$po_file"
+        
+        # If a new file has been added to /wiki/en/, add message after sed error to clarify it will be created
+        if ! [ -f "$po_file" ] ; then
+            echo creating "$po_file"
+        fi
 
         if ! po4a-updatepo \
             --format asciidoc \
@@ -79,15 +61,19 @@ do
             --po "$po_file" ; then
         echo ''
         echo Error updating "$lang" PO file for: "$adoc_file"
-
         fi
     done
-
 done <   <(find -L "$SRC_DIR" -name "*.md" -print0)
 
 echo ''
-echo 'REMOVE TEMPORARY FILES'
+echo Removing temporary files
 
-for lang in $(ls "$PO_DIR" ) ; do
-	rm "$PO_DIR/$lang/"*.po~
+for lang in $(ls "$PO_DIR") ; do
+
+    # Check if the temporary files exist before removing them to prevent misleading error message
+    temp_file="$PO_DIR/$lang/*.po~"
+
+    if ls $temp_file 1> /dev/null 2>&1 ; then
+	    rm "$PO_DIR/$lang/"*.po~
+    fi
 done
